[ADD] #68 - MyPageSettingActivity.kt, basic view setting
