+++ app/src/main/java/org/mydaily/ui/view/MainActivity.kt
@@ -28,7 +28,6 @@ class MainActivity : BaseActivity<ActivityMainBinding, UserViewModel>() {
     override fun initView() {
         initToolbar()
         initBottomNavigation()
-
         replaceDailyFragment()
     }
