+++ app/src/main/java/org/mydaily/ui/view/keyword/KeywordSelectActivity.kt
@@ -2,8 +2,10 @@ package org.mydaily.ui.view.keyword
 
 import android.app.AlertDialog
 import android.content.Intent
+import android.graphics.drawable.ColorDrawable
 import android.view.Menu
 import android.view.MenuItem
+import androidx.core.content.ContextCompat
 import com.google.android.material.chip.Chip
 import com.google.android.material.chip.ChipDrawable
 import org.koin.androidx.viewmodel.ext.android.viewModel
@@ -41,9 +43,8 @@ class KeywordSelectActivity : BaseActivity<ActivityKeywordSelectBinding, Keyword
 
     private fun initToolbar() {
         setSupportActionBar(binding.tbKeywordSelectActivity)
-        supportActionBar?.setDisplayHomeAsUpEnabled(true)
         supportActionBar?.setDisplayShowTitleEnabled(false)
-        supportActionBar?.setHomeAsUpIndicator(R.drawable.ic_arrow_back)
+        supportActionBar?.setBackgroundDrawable(ColorDrawable(ContextCompat.getColor(this, R.color.white)))
         binding.tbKeywordSelectActivity.setNavigationOnClickListener {
             finish()
         }
@@ -57,9 +58,9 @@ class KeywordSelectActivity : BaseActivity<ActivityKeywordSelectBinding, Keyword
 
     private fun floatingDialog() {
         AlertDialog.Builder(this)
-            .setTitle(R.string.up_to_eight)
-            .setMessage(R.string.too_many_keyword_selected)
-            .setPositiveButton("확인", null)
+            .setTitle(R.string.up_to_four)
+            .setMessage(R.string.think_more)
+            .setPositiveButton(getString(R.string.okay), null)
             .create()
             .show()
     }
