+++ app/src/main/res/drawable/ic_help_orange.xml
@@ -1,0 +1,22 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="24dp"
+    android:height="24dp"
+    android:viewportWidth="24"
+    android:viewportHeight="24">
+  <path
+      android:strokeWidth="1"
+      android:pathData="M12,14C12,12 13.576,11.6653 14.1187,11.124C14.6631,10.5809 15,9.8298 15,9C15,7.3432 13.6569,6 12,6C11.1041,6 10.2999,6.3927 9.7502,7.0154C9.496,7.3033 9.2963,7.6404 9.167,8.0106"
+      android:fillColor="#00000000"
+      android:strokeColor="#FF7556"
+      android:strokeLineCap="round"/>
+  <path
+      android:pathData="M12,18C12.5523,18 13,17.5523 13,17C13,16.4477 12.5523,16 12,16C11.4477,16 11,16.4477 11,17C11,17.5523 11.4477,18 12,18Z"
+      android:fillColor="#FF7556"
+      android:fillType="evenOdd"/>
+  <path
+      android:strokeWidth="1"
+      android:pathData="M12,22C17.5228,22 22,17.5228 22,12C22,6.4771 17.5228,2 12,2C6.4771,2 2,6.4771 2,12C2,17.5228 6.4771,22 12,22Z"
+      android:fillColor="#00000000"
+      android:fillType="evenOdd"
+      android:strokeColor="#FF7556"/>
+</vector>
