+++ app/src/main/res/layout/activity_keyword_list.xml
@@ -1,5 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <layout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools"
     xmlns:app="http://schemas.android.com/apk/res-auto">
 
     <data>
@@ -8,16 +9,18 @@
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
-        android:layout_height="match_parent">
+        android:layout_height="match_parent"
+        tools:context=".ui.view.keyword.KeywordListActivity">
 
         <androidx.appcompat.widget.Toolbar
-            android:id="@+id/tb_keyword_add_activity"
+            android:id="@+id/tb_keyword_list_activity"
             android:layout_width="match_parent"
             android:layout_height="?attr/actionBarSize"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintTop_toTopOf="parent"
-            app:menu="@menu/menu_keyword_list">
+            app:menu="@menu/menu_keyword_list"
+            app:navigationIcon="@drawable/ic_btn_arrow_left">
 
             <TextView
                 android:layout_width="wrap_content"
@@ -30,7 +33,6 @@
 
         </androidx.appcompat.widget.Toolbar>
 
-
         <ScrollView
             android:id="@+id/sv_keyword_list"
             android:layout_width="match_parent"
@@ -38,7 +40,7 @@
             app:layout_constraintBottom_toTopOf="@+id/btn_select_finish"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toBottomOf="@+id/tb_keyword_add_activity">
+            app:layout_constraintTop_toBottomOf="@+id/tb_keyword_list_activity">
 
             <androidx.constraintlayout.widget.ConstraintLayout
                 android:id="@+id/cl_in_sv"
@@ -84,21 +86,21 @@
 
                 <com.google.android.material.chip.ChipGroup
                     android:id="@+id/cg_life"
+                    style="@style/Widget.MaterialComponents.Chip.Choice"
                     android:layout_width="match_parent"
                     android:layout_height="wrap_content"
-                    android:paddingHorizontal="16dp"
                     android:layout_marginTop="16dp"
-                    style="@style/Widget.MaterialComponents.Chip.Choice"
-                    app:layout_constraintTop_toBottomOf="@id/tv_life"
+                    android:paddingHorizontal="16dp"
+                    app:layout_constraintEnd_toEndOf="parent"
                     app:layout_constraintStart_toStartOf="parent"
-                    app:layout_constraintEnd_toEndOf="parent"/>
+                    app:layout_constraintTop_toBottomOf="@id/tv_life" />
 
                 <TextView
                     android:id="@+id/tv_work"
                     android:layout_width="wrap_content"
                     android:layout_height="wrap_content"
-                    android:layout_marginTop="48dp"
                     android:layout_marginStart="16dp"
+                    android:layout_marginTop="48dp"
                     android:text="@string/attitude_for_work"
                     android:textColor="@color/cod_gray"
                     android:textSize="16sp"
@@ -110,18 +112,18 @@
                     android:id="@+id/cg_work"
                     android:layout_width="match_parent"
                     android:layout_height="wrap_content"
-                    android:paddingHorizontal="16dp"
                     android:layout_marginTop="16dp"
-                    app:layout_constraintTop_toBottomOf="@id/tv_work"
+                    android:paddingHorizontal="16dp"
+                    app:layout_constraintEnd_toEndOf="parent"
                     app:layout_constraintStart_toStartOf="parent"
-                    app:layout_constraintEnd_toEndOf="parent"/>
+                    app:layout_constraintTop_toBottomOf="@id/tv_work" />
 
                 <TextView
                     android:id="@+id/tv_my_word"
                     android:layout_width="wrap_content"
                     android:layout_height="wrap_content"
-                    android:layout_marginTop="48dp"
                     android:layout_marginStart="16dp"
+                    android:layout_marginTop="48dp"
                     android:text="@string/my_word"
                     android:textColor="@color/cod_gray"
                     android:textSize="16sp"
@@ -154,21 +156,21 @@
                     android:id="@+id/cg_my_word"
                     android:layout_width="match_parent"
                     android:layout_height="wrap_content"
-                    android:paddingHorizontal="16dp"
                     android:layout_marginTop="16dp"
-                    app:layout_constraintTop_toBottomOf="@id/tv_my_word"
+                    android:paddingHorizontal="16dp"
+                    app:layout_constraintEnd_toEndOf="parent"
                     app:layout_constraintStart_toStartOf="parent"
-                    app:layout_constraintEnd_toEndOf="parent">
+                    app:layout_constraintTop_toBottomOf="@id/tv_my_word">
 
                     <com.google.android.material.chip.Chip
                         android:id="@+id/chip_add"
                         android:layout_width="wrap_content"
                         android:layout_height="wrap_content"
+                        android:stateListAnimator="@null"
+                        app:chipBackgroundColor="@android:color/transparent"
                         app:chipIcon="@drawable/ic_add"
                         app:chipIconSize="32dp"
-                        app:rippleColor="@android:color/transparent"
-                        android:stateListAnimator="@null"
-                        app:chipBackgroundColor="@android:color/transparent"/>
+                        app:rippleColor="@android:color/transparent" />
 
                 </com.google.android.material.chip.ChipGroup>
