+++ app/src/main/res/menu/menu_keyword_list.xml
@@ -4,8 +4,8 @@
 
     <item
         android:id="@+id/menu_help"
-        android:icon="@drawable/ic_help"
+        android:icon="@drawable/ic_help_orange"
         android:title="Help"
-        app:showAsAction="ifRoom"/>
+        app:showAsAction="ifRoom" />
 
 </menu>
